pyproject.toml
[tool.poetry]
name = "RandomBot"
version = "0.1"
description = "My Discord bot."
authors = ["GD Tom <hairy.furball154@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.15.0"
discord = "^1.0.1"
flask = "^1.1.2"
replit = "^2.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
